requirements.txt
fastapi>=0.63.0
pydantic>=1.8.1, <2.0.0
tortoise-orm>=0.16.21
pytest>=6.2.2
